Refactor random number handling and data types in coinflip program
- Updated the `random_number` parameter type in the `SetOracleHash` instruction from `usize` to `u64`, improving consistency in data types.
- Simplified the calculation of the final number in the `Game` struct by removing the game ID from the computation, enhancing clarity in winner determination.

These changes improve the maintainability and clarity of the coinflip program by refining data type usage and simplifying logic in game state management.

+++ programs/coinflip/src/instructions/accounts.rs
@@ -262,7 +262,7 @@ pub struct JoinGame<'info> {
 
 #[derive(Accounts)]
 #[instruction(
-    random_number: usize,
+    random_number: u64,
 )]
 pub struct SetOracleHash<'info> {
     #[account(
@@ -278,7 +278,7 @@ pub struct SetOracleHash<'info> {
         constraint = game_token.token_mint == game.token_mint @ ErrorCode::InvalidToken,
     )]
     #[account(
-        seeds = [b"player_token", game.calculate_winner(random_number, Clock::get()?.unix_timestamp as u64).as_ref(), game_token.token_mint.as_ref()],
+        seeds = [b"player_token", game.calculate_winner(random_number as usize, Clock::get()?.unix_timestamp as u64).as_ref(), game_token.token_mint.as_ref()],
         bump,
     )]
     pub player_token: Account<'info, PlayerToken>,
